+++ Sources/Informant/TtyTextFormatter.swift
@@ -54,6 +54,8 @@ private extension DocProblem.Detail {
             return "Parameters are organized in the \("separate", color: .green) style, but \("grouped", color: .green) style is preferred."
         case .parametersAreNotSeparated:
             return "Parameters are organized in the \("grouped", color: .green) style, but \("separate", color: .green) style is preferred."
+        case .missingColon(let name):
+            return "\(name, color: .green) should be followed by a \(":", color: .green) character."
         }
     }